Cleaned up list a little bit, also cleaned up the code
Biggest things I did was fix receive so that directory contents only
show once(this will also help a lot with recieving files). I also
retooled LIST a little so it uses the pointer to returned by recv
command.

I also deleted about 100 lines of code we wont be using. If for some
reason someone needs it, it will be in the subversion.