Got changeWorkingDirectory working or the 'cd' command..
also got getFile started.. works sometimes, and sometimes not. really bugging me so im going to let Kyle take a look at it.

also LS fails to work if the directory is empty.

I think we will have to either end up forking or using a thread for the commands that take PASV, because of the weird errors sometimes, and sometimes it works...

but im tired. so im done for now.